msm: Make the pmem SMI request be not active only
The pmem bandwidth voting request is supposed to be a failsafe
fallback. If active_only is set, the bandwidth request will
be dropped if apps goes into idle power collapse, which could
potentially cause a bus lockup. Make sure the fail safe is
actually safe by making the request not active_only.

Change-Id: Iced1dfdd1fb325529ac354624e3ecf38206bb7b5
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>